feature: Improve ui on drop off

+++ features/drop-off/src/main/java/com/carissa/revibes/drop_off/data/DropOffRepository.kt
@@ -31,7 +31,7 @@ data class SubmitOrderItemData(
     val id: String,
     val name: String,
     val type: String,
-    val weight: Double
+    val weight: Int
 )
 
 @Single
