+++ features/drop-off/src/main/java/com/carissa/revibes/drop_off/data/model/SubmitOrderResponse.kt
@@ -27,5 +27,5 @@ data class SubmitOrderItem(
     val id: String,
     val name: String,
     val type: String,
-    val weight: Double
+    val weight: Int
 )
